# Starlark: optimze list(list)