--- a/src/main/java/net/starlark/java/eval/StarlarkList.java
+++ b/src/main/java/net/starlark/java/eval/StarlarkList.java
@@ -535,4 +535,16 @@ public final class StarlarkList<E> extends AbstractList<E>
   public Object[] toArray() {
     return size != 0 ? Arrays.copyOf(elems, size, Object[].class) : EMPTY_ARRAY;
   }
+
+  @SuppressWarnings("unchecked")
+  @Override
+  public <T> T[] toArray(T[] a) {
+    if (a.length < size) {
+      return (T[]) Arrays.copyOf(elems, size, a.getClass());
+    } else {
+      System.arraycopy(elems, 0, a, 0, size);
+      Arrays.fill(a, size, a.length, null);
+      return a;
+    }
+  }
 }