--- a/src/test/java/net/starlark/java/eval/BUILD
+++ b/src/test/java/net/starlark/java/eval/BUILD
@@ -22,6 +22,7 @@ java_test(
         "MethodLibraryTest.java",
         "MutabilityTest.java",
         "PrinterTest.java",
+        "TupleTest.java",
         "StarlarkAnnotationsTest.java",
         "StarlarkEvaluationTest.java",
         "StarlarkFlagGuardingTest.java",