--- a/src/test/java/net/starlark/java/eval/EvalTests.java
+++ b/src/test/java/net/starlark/java/eval/EvalTests.java
@@ -32,5 +32,6 @@ import org.junit.runners.Suite;
   StarlarkMutableTest.class,
   StarlarkThreadDebuggingTest.class,
   StarlarkThreadTest.class,
+  TupleTest.class,
 })
 public class EvalTests {}