--- a/src/test/java/net/starlark/java/eval/testdata/bench_list.star
+++ b/src/test/java/net/starlark/java/eval/testdata/bench_list.star
@@ -1,4 +1,5 @@
 _list1000 = list(range(1000))
+_list10 = list(range(10))
 
 def bench_extend(b):
     "Extends an empty list by 1000 items."
@@ -12,3 +13,8 @@ def bench_append(b):
         x = []
         for y in _list1000:
             x.append(y)
+
+def bench_list_list(b):
+    "list([10 elements])"
+    for _ in range(b.n):
+        list(_list10)