`toArray()` is more efficient than `toArray(new Object[size()])`,
so switch to `toArray()` in `Starlark.toArray(..)`.

```
A: N=23 r=8.124+-0.325 se=0.069
B: N=23 r=6.613+-0.391 se=0.083
B/A: 0.814
```

for a benchmark

```
l = [1, 2, 3, 4, 5, 6, 7, 8, 9, 10]
for i in range(50000000):
    list(l)
```

Or for bundled benchmark:

```
benchmark                   ops     cpu/op    wall/op   steps/op   alloc/op
# before
bench_list_list        67108863      209ns      208ns          4       255B
# after
bench_list_list        67108863      174ns      173ns          4       223B
```