@alandonovan 2020-11-26
> toArray() is more efficient than toArray(new Object[size()])

Perhaps StarlarkList and Tuple should override toArray(Object[]).