@stepancheg 2020-11-26
* added `toArray(Object[])` specialisations in `Tuple` and `StarlarkList` (unused, but added tests)
* added `Sequence.toArray()` override to clarify it must not return `Object[]` subclasses

Note [Collection.toArray() javadoc](https://docs.oracle.com/en/java/javase/11/docs/api/java.base/java/util/Collection.html#toArray()) is self-contradictory or at least confusing:

> The returned array will be "safe" in that no references to it are maintained by this collection. (In other words, this method must allocate a new array even if this collection is backed by an array). The caller is thus free to modify the returned array.

It is safe to modify returned array if returned array is empty. In Starlark implementations of `toArray()` we return a preallocated empty instance from empty collections. But javadoc says the implementation *must* allocate a new array.

Perhaps, if we want to be strict about following the spec, we should add another `Sequence` operation, like `Sequence.starlarkToArray()` with semantics we need.