@stepancheg 2020-11-29
Reverted implementation of `Starlark.toArray()`, kept only specialization to `toArray(Object[])` in `Tuple` and `StarlarkList`.